Remove the rest of the mapper.
