+++ lib/encoders/encoders.h
@@ -14,7 +14,6 @@ public:
     virtual ~AbstractEncoder() {}
 
     static std::unique_ptr<AbstractEncoder> create(const EncoderProto& config);
-
 public:
     virtual std::vector<std::shared_ptr<const Sector>> collectSectors(
         const Location& location, const Image& image) = 0;
@@ -22,6 +21,9 @@ public:
     virtual std::unique_ptr<Fluxmap> encode(const Location& location,
         const std::vector<std::shared_ptr<const Sector>>& sectors,
         const Image& image) = 0;
+
+    nanoseconds_t calculatePhysicalClockPeriod(
+        nanoseconds_t targetClockPeriod, nanoseconds_t targetRotationalPeriod);
 };
 
 #endif
